refactor: move sanitize_diff_args function to utils for better organization
- Moved the `sanitize_diff_args` function from `cli.py` to `utils.py` to enhance code organization and maintainability.
- Updated imports in `cli.py` to reflect the new location of the function, ensuring continued functionality.
