+++ diffgraph/cli.py
@@ -8,6 +8,7 @@ import os
 from diffgraph.ai_analysis import CodeAnalysisAgent
 from diffgraph.html_report import generate_html_report, AnalysisResult
 from diffgraph.env_loader import load_env_file, debug_environment
+from diffgraph.utils import sanitize_diff_args
 
 # Load environment variables
 load_env_file()
@@ -108,79 +109,6 @@ def get_changed_files(diff_args: List[str] = None) -> List[Dict[str, str]]:
             sys.exit(1)
 
     return changed_files
-
-def sanitize_diff_args(diff_args: List[str]) -> List[str]:
-    """
-    Sanitize diff arguments to prevent command injection and ensure safe execution.
-
-    Args:
-        diff_args: List of diff arguments to sanitize
-
-    Returns:
-        List of sanitized, safe diff arguments
-    """
-    if not diff_args:
-        return []
-
-    # Dangerous flags that could cause issues or suppress patch content
-    dangerous_flags = {
-        '--name', '--name-only', '--name-status', '--pretty', '--numstat',
-        '--format', '--exec', '--output-format', '--color',
-        '--no-color', '--color=always', '--color=auto', '--color=never'
-    }
-
-    # Safe flags that are commonly needed
-    safe_flags = {
-        '-U', '--unified', '-R', '--reverse', '-B', '--break-rewrites',
-        '-M', '--find-renames', '-C', '--find-copies', '--find-copies-harder',
-        '-D', '--irreversible-delete', '-l', '--max-count', '-S', '--find-object',
-        '-G', '--pickaxe-regex', '--pickaxe-all', '--pickaxe-regex',
-        '--relative', '--no-relative', '--text', '--ignore-space-at-eol',
-        '--ignore-space-change', '--ignore-all-space', '--ignore-blank-lines',
-        '--indent-heuristic', '--patience', '--histogram', '--minimal',
-        '--anchored', '--word-diff', '--word-diff-regex', '--color-words',
-        '--no-renames', '--check', '--ws-error-highlight', '--full-index',
-        '--binary', '--abbrev', '--src-prefix', '--dst-prefix', '--no-prefix'
-    }
-
-    sanitized = []
-    blocked_flags = []
-
-    for arg in diff_args:
-        # Block clearly dangerous flags with clear communication
-        if arg in dangerous_flags:
-            blocked_flags.append(arg)
-            continue
-
-        # Allow safe flags
-        if arg in safe_flags:
-            sanitized.append(arg)
-            continue
-
-        # Allow commit references (SHA, branch names, etc.)
-        if not arg.startswith('-'):
-            sanitized.append(arg)
-            continue
-
-        # Allow numeric values for context lines
-        if arg.startswith('-') and arg[1:].isdigit():
-            sanitized.append(arg)
-            continue
-
-        # For unknown flags, trust the user but warn them
-        click.secho(f"⚠️  Warning: Unknown diff argument '{arg}' - allowing but use with caution", fg="yellow")
-        sanitized.append(arg)
-
-    # Always add --no-color for consistent, parseable output
-    if '--no-color' not in sanitized:
-        sanitized.append('--no-color')
-
-    # Report any blocked dangerous flags
-    if blocked_flags:
-        click.secho(f"🚫 Blocked dangerous diff arguments: {', '.join(blocked_flags)}", fg="red")
-        click.secho("   These flags could cause security issues or suppress patch content", fg="red")
-
-    return sanitized
 
 def load_file_contents(changed_files: List[Dict[str, str]], diff_args: List[str] = None) -> List[Dict[str, str]]:
     """
