+++ diffgraph/utils/__init__.py
@@ -1,0 +1,7 @@
+"""
+Utility functions for the DiffGraph package.
+"""
+
+from .git_utils import sanitize_diff_args
+
+__all__ = ['sanitize_diff_args']
